+++ src/Opulence/Debug/Exceptions/Handlers/Http/templates/json/MaintenanceMode.php
@@ -1,0 +1,7 @@
+<?php
+echo json_encode([
+    "error" => [
+        "code" => htmlentities($statusCode, ENT_QUOTES, "UTF-8"),
+        "message" => "Please check back later"
+    ]
+]);