+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * Renamed `build/commit.sh` to `build/git.sh`
 * Added checking out and merging of pull requests
 
+<h3>Debug</h3>
+* Added support for different response formats depending on the request type
+
 <h3>HTTP</h3>
 * Added support for trusted proxies
 * Removed all `Request::METHOD_*` constants, moved them to `Opulence\Http\Requests\RequestMethods` class, and dropped `METHOD_` from constants' names